Tradingview/options-dashboard/requirements.txt
streamlit>=1.29.0
plotly>=5.18.0
alpaca-py>=0.13.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
